- Remove Unused Functions - Refactored helper function
